MAC and IP address
Questions:

What is a MAC address?

The name of a network interface
The unique identifier of a network interface
A network interface

What is an IP address?

Is to devices connected to a network what postal address is to houses
The unique identifier of a network interface
Is a number that network devices use to connect to networks